refactor: moved test to own location
