+++ tests/test_unit/test_cdf_tk/test_utils/test_utils.py
@@ -20,7 +20,6 @@ from cognite.client.data_classes.capabilities import (
     ProjectCapabilityList,
     ProjectsScope,
 )
-from cognite.client.data_classes.data_modeling import DataModelId, ViewId
 from cognite.client.data_classes.iam import ProjectSpec
 from cognite.client.exceptions import CogniteAuthError
 from pytest import MonkeyPatch
@@ -32,7 +31,6 @@ from cognite_toolkit._cdf_tk.utils import (
     AuthReader,
     AuthVariables,
     CDFToolConfig,
-    GraphQLParser,
     calculate_directory_hash,
     flatten_dict,
     iterate_modules,
@@ -442,221 +440,6 @@ def test_flatten_dict(input_: dict[str, Any], expected: dict[str, Any]) -> None:
     actual = flatten_dict(input_)
 
     assert actual == expected
-
-
-SPACE = "sp_my_space"
-DATA_MODEL = DataModelId(SPACE, "MyDataModel", "v1")
-GraphQLTestCases = [
-    pytest.param(
-        """interface Creatable @view(space: "cdf_apps_shared", version: "v1") @import {
-  visibility: String
-  createdBy: CDF_User
-  updatedBy: CDF_User
-  isArchived: Boolean
-}""",
-        DATA_MODEL,
-        set(),
-        {ViewId("cdf_apps_shared", "Creatable", "v1")},
-        id="Imported interface",
-    ),
-    pytest.param(
-        """"@code WOOL"
-type WorkOrderObjectListItem @import(dataModel: {externalId: "MaintenanceDOM", version: "2_2_0", space: "EDG-COR-ALL-DMD"}) {
-  ...
-}""",
-        DATA_MODEL,
-        set(),
-        {DataModelId(space="EDG-COR-ALL-DMD", external_id="MaintenanceDOM", version="2_2_0")},
-        id="Imported data model",
-    ),
-    pytest.param(
-        '''
-"""
-@name        Notification
-@description This is a SAP notification object.
-@code        NOTI
-"""
-type Notification @container(indexes: [{identifier: "refCostCenterIndex", indexType: BTREE, fields: ["refCostCenter"]}, {identifier: "createdDateIndex", indexType: BTREE, fields: ["createdDate"], cursorable: true}]) @import(dataModel: {externalId: "MaintenanceDOM", version: "2_2_0", space: "EDG-COR-ALL-DMD"}) {
-  ...
-  }
-''',
-        DATA_MODEL,
-        set(),
-        {DataModelId(space="EDG-COR-ALL-DMD", external_id="MaintenanceDOM", version="2_2_0")},
-        id="Imported data model with unused container directive",
-    ),
-    pytest.param(
-        '''
-    """
-@code WKCC
-@Description Work Center Category in SAP
-"""
-type WorkCenterCategory {
-    """
-    @name Name
-    """
-    name: String!
-    """
-    @name Description
-    """
-    description: String
-    """
-    @name Code
-    """
-    code: String
-}''',
-        DATA_MODEL,
-        {ViewId(SPACE, "WorkCenterCategory", None)},
-        set(),
-        id="Simple type",
-    ),
-    pytest.param(
-        """type Cdf3dConnectionProperties
-  @import
-  @view(space: "cdf_3d_schema", version: "1")
-  @edge
-  @container(
-    constraints: [
-      {
-        identifier: "uniqueNodeRevisionConstraint"
-        constraintType: UNIQUENESS
-        fields: ["revisionId", "revisionNodeId"]
-      }
-    ]
-  ) {
-  revisionId: Int64!
-  revisionNodeId: Int64!
-}""",
-        DATA_MODEL,
-        set(),
-        {ViewId("cdf_3d_schema", "Cdf3dConnectionProperties", "1")},
-        id="Edge type",
-    ),
-    pytest.param(
-        """type APM_User @view (version: "7") {
-  name: String
-  email: String
-  lastSeen: Timestamp
-  preferences: JSONObject
-}""",
-        DATA_MODEL,
-        {ViewId(SPACE, "APM_User", "7")},
-        set(),
-        id="Simple type with version",
-    ),
-    pytest.param(
-        """type UnitOfMeasurement
-  @import(
-    dataModel: {
-      externalId: "CoreDOM"
-      version: "1_0_18"
-      space: "EDG-COR-ALL-DMD"
-    }
-  ) {
-  name: String!
-}""",
-        DATA_MODEL,
-        set(),
-        {DataModelId(space="EDG-COR-ALL-DMD", external_id="CoreDOM", version="1_0_18")},
-        id="No comma, only newline",
-    ),
-    pytest.param(
-        '''"""
-  @name Name }{ Breaks the parser
-"""
-type UnitOfMeasurement
-  @import(
-    dataModel: {
-      externalId: "CoreDOM"
-      version: "1_0_18"
-      space: "EDG-COR-ALL-DMD"
-    }
-  ) {
-  name: String!
-}''',
-        DATA_MODEL,
-        set(),
-        {DataModelId(space="EDG-COR-ALL-DMD", external_id="CoreDOM", version="1_0_18")},
-        id="Ignore comments",
-    ),
-    pytest.param(
-        """type APM_Config @view {
-  name: String
-  appDataSpaceId: String
-  appDataSpaceVersion: String
-  customerDataSpaceId: String
-  customerDataSpaceVersion: String
-  featureConfiguration: JSONObject
-  fieldConfiguration: JSONObject
-  rootLocationsConfiguration: JSONObject
-}""",
-        DATA_MODEL,
-        {ViewId(SPACE, "APM_Config", None)},
-        set(),
-        id="No version",
-    ),
-    pytest.param(
-        """"Navigational aid for traversing CogniteCADModel instances"
-type CogniteCADModel implements CogniteDescribable & Cognite3DModel
-  @view(
-    space: "cdf_cdm"
-    version: "v1"
-    rawFilter: {
-      and: [
-        {
-          hasData: [
-            {
-              type: "container"
-              space: "cdf_cdm_3d"
-              externalId: "Cognite3DModel"
-            }
-          ]
-        }
-        {
-          equals: {
-            property: ["cdf_cdm_3d", "Cognite3DModel", "type"]
-            value: "CAD"
-          }
-        }
-      ]
-    }
-  )
-  @import {
-  "Name of the instance"
-  name: String
-  "Description of the instance"
-  description: String
-  "Text based labels for generic use, limited to 1000"
-  tags: [String]
-  "Alternative names for the node"
-  aliases: [String]
-  "CAD, PointCloud or Image360"
-  type: Cognite3DModel_type
-  "Thumbnail of the 3D model"
-  thumbnail: CogniteFile
-  "List of revisions for this CAD model"
-  revisions: [CogniteCADRevision]
-    @reverseDirectRelation(throughProperty: "model3D")
-}""",
-        DATA_MODEL,
-        set(),
-        {ViewId("cdf_cdm", "CogniteCADModel", "v1")},
-        id="Setting custom filter on view",
-    ),
-]
-
-
-class TestGraphQLParser:
-    @pytest.mark.parametrize("raw, data_model_id, expected_views, dependencies", GraphQLTestCases)
-    def test_parse(
-        self, raw: str, data_model_id: DataModelId, expected_views: set[ViewId], dependencies: set[ViewId | DataModelId]
-    ) -> None:
-        parser = GraphQLParser(raw, data_model_id)
-
-        actual_views = parser.get_views(include_version=True)
-        assert expected_views == actual_views
-        actual_dependencies = parser.get_dependencies(include_version=True)
-        assert dependencies == actual_dependencies
 
 
 def quote_key_in_yaml_test_cases() -> Iterable[ParameterSet]:
